requirements.txt
Flask-CKEditor==0.4.4.1
flask==2.2.2
flask-sqlalchemy==3.0.2
WTForms~=3.0.1
Flask-Login==0.6.2
Werkzeug==2.2.3
Flask-Gravatar==0.5.0
SQLAlchemy==2.0.3
gunicorn==20.1.0
python-decouple==3.8
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.1.1
email-validator==1.3.1
psycopg2-binary==2.9.5